Добавляет данные в форму создания формы
